imprived checking of existing program versions and made vm.max_map_count persistent

+++ install.sh
@@ -31,17 +31,31 @@ else
 fi
 
 # Docker Compose
-wget https://github.com/docker/compose/releases/download/v2.2.3/docker-compose-linux-x86_64 -O /usr/local/bin/docker-compose
-chmod +x /usr/local/bin/docker-compose
-printf '\nDocker Compose installed successfully\n\n'
-#
-# Copy docker daemon conf file
-#
-cp daemon.json /etc/docker/
+currentver="$(docker-compose --version | cut -d' ' -f4)"
+requiredver="2.2.3"
+if [ "$(printf '%s\n' "$requiredver" "$currentver" | sort -V | head -n1)" = "$requiredver" ]; then 
+	echo "Docker-Compose Greater than or equal to ${requiredver}"
+else
+	echo "Docker-Compose Less than ${requiredver}"
+	wget https://github.com/docker/compose/releases/download/v2.2.3/docker-compose-linux-x86_64 -O /usr/local/bin/docker-compose
+	chmod +x /usr/local/bin/docker-compose
+	printf '\nDocker Compose installed successfully\n\n'
+	#
+	# Copy docker daemon conf file
+	#
+	cp daemon.json /etc/docker/
+fi
 
 
 printf 'Launch RAMP IoT platform containers...\n\n'
 export RAMP_PATH=/opt/rampiot
+
+
+if grep -q "vm.max_map_count=262144" /etc/sysctl.conf; then
+	echo "vm.max_map_count sufficient"
+else
+	echo "vm.max_map_count=262144" >> /etc/sysctl.conf
+fi
 
 sudo sysctl -w vm.max_map_count=262144
 sudo docker-compose up -d
@@ -50,4 +64,10 @@ sudo docker-compose up -d
 #
 # Copy desktop launcher
 #
-sudo ln -s /opt/rampiot/rampiot.desktop /usr/share/applications/
+cp rampiot.desktop /opt/rampiot/
+FILE=/usr/share/applications/rampiot.desktop
+if test -f "$FILE"; then
+	echo "$FILE exists."
+else
+	sudo ln -s /opt/rampiot/rampiot.desktop /usr/share/applications/
+fi
